scripts/requirements.txt
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
databricks-sdk==0.61.0
google-auth==2.40.3
idna==3.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
requests==2.32.4
rsa==4.9.1
urllib3==2.5.0
